requirements.txt
pandas
joblib
numpy
scipy
scikit-learn==1.2.2
tensorflow
keras
protobuf
Pillow
streamlit==1.26
plotly==5.16.0
click==8.0.4
s3fs
st-files-connection
boto3
hydralit